pyproject.toml
[project]
name = "shorts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "lxml>=6.0.2",
    "openai>=1.108.2",
    "requests>=2.32.5",
    "python-dotenv>=1.0.1",
    "google>=3.0.0",
    "google-genai>=1.38.0",
    "moviepy>=2.2.1",
]
